JCF: Issue #20: determine the modules owned by each application
More broadly, this is the first commit in an overhaul of how to plot
the configuration. The primary change is that now the
generate_modules() functions defined in appmodel are used to create
modules which can be used as vertices in the plot. Having said that, a
few things to note:

-std::cout is used on this feature branch for ease of
 reading-to-screen, and C++17 is used for legacy reasons; this will
 not be the case when the PR comes in

-Still to-do is a calculation of the network connections between
 applications and queue connections between modules

-Also to-do is the physical construction of the graph; for now just
 printing to screen what's happening